#4415 allows the devops-edit command to update the build config and any related webhooks if used from the CLI or inside eclipse etc
